+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # TBD
 
+### Features
+- Add `wait` and `define` command in Startosis
+
+### Fixes
+- Fixes a bug where we'd propagate a nil error
+
 # 0.52.0
 ### Breaking Changes
 - Unified `GetUserServiceLogs` and `StreamUserServiceLogs` engine's endpoints, now `GetUserServiceLogs` will handle both use cases